Simulate movement tracks

Wrapper function to simulate tracks under multiple scenarios/movement models.

Usage

simulate_tracks(
  scenarios,
  n_steps,
  n_tracks,
  start_loc,
  sl_par,
  prange = NA,
  jitter = TRUE,
  neighbors = NA
)

Arguments

scenarios

Data frame defining parameter values for each scenario as formatted by create_scenarios_crw(), create_scenarios_ccrw(), etc.

n_steps

Numeric. Number of steps to simulate

n_tracks

Numeric. Number of tracks to simulate.

start_loc

Data frame with a single row containing coordinates of the starting location (columns x and y). 0, 0 if unprojected.

sl_par

Vector of length 2 containing the shape and scale of the step length distribution (Weibull)

prange

Numeric. Perceptual range radius. Units match the units of the coordinates: m if UTM, degrees if lat/long.

neighbors

List of cell neighborhoods created by get_neighbors. #' @param jitter Logical. If TRUE, starting locations will be jittered by 20 km in any direction for spatially-explicit models (MCRW and BCRW).

Value

Returns a data frame containing the coordinates of each location along the simulated tracks and the associated simulation parameter values as specified in scenarios. Each track is assigned an individual id

under each scenario.